requirements.txt
aiogram==2.25.1
Pillow==10.2.0
requests==2.31.0
